split the hugh python file
